+++ packages/devtools_app/test/deep_link_vlidation/deep_links_screen_test.dart
@@ -5,6 +5,7 @@
 import 'package:devtools_app/devtools_app.dart';
 import 'package:devtools_app/src/screens/deep_link_validation/deep_link_list_view.dart';
 import 'package:devtools_app/src/screens/deep_link_validation/deep_links_model.dart';
+import 'package:devtools_app/src/screens/deep_link_validation/validation_details_view.dart';
 import 'package:devtools_app/src/shared/directory_picker.dart';
 import 'package:devtools_app_shared/ui.dart';
 import 'package:devtools_app_shared/utils.dart';
@@ -40,14 +41,14 @@ void main() {
       ),
     );
     deferredLoadingSupportEnabled = true;
-    await tester.pumpAndSettle(const Duration(seconds: 1));
+    await tester.pump(const Duration(seconds: 1));
     expect(find.byType(DeepLinkPage), findsOneWidget);
   }
 
   group('DeepLinkScreen', () {
     setUp(() {
       screen = DeepLinksScreen();
-      deepLinksController = DeepLinksController();
+      deepLinksController = DeepLinksTestController();
     });
 
     testWidgets('builds its tab', (WidgetTester tester) async {
@@ -75,18 +76,11 @@ void main() {
     );
 
     testWidgetsWithWindowSize(
-      'builds deeplink list page',
+      'builds deeplink list page with no links',
       windowSize,
       (WidgetTester tester) async {
         deepLinksController.selectedProject.value =
             FlutterProject(path: '/abc', androidVariants: ['debug', 'release']);
-        deepLinksController.linkDatasNotifier.value = [
-          LinkData(
-            domain: 'www.google.com',
-            path: '/',
-            os: [PlatformOS.android],
-          ),
-        ];
         await pumpDeepLinkScreen(
           tester,
           controller: deepLinksController,
@@ -94,7 +88,45 @@ void main() {
 
         expect(find.byType(DeepLinkPage), findsOneWidget);
         expect(find.byType(DeepLinkListView), findsOneWidget);
+        expect(find.byType(CenteredCircularProgressIndicator), findsOneWidget);
+      },
+    );
+
+    testWidgetsWithWindowSize(
+      'builds deeplink list page with links and split screen',
+      windowSize,
+      (WidgetTester tester) async {
+        final deepLinksController = DeepLinksTestController();
+
+        deepLinksController.selectedProject.value =
+            FlutterProject(path: '/abc', androidVariants: ['debug', 'release']);
+        final linkData = LinkData(
+          domain: 'www.google.com',
+          path: '/',
+          os: [PlatformOS.android],
+        );
+        deepLinksController.allLinkDatasNotifier.value = [linkData];
+        deepLinksController.displayOptionsNotifier.value =
+            DisplayOptions(showSplitScreen: true);
+        deepLinksController.selectedLink.value = linkData;
+
+        await pumpDeepLinkScreen(
+          tester,
+          controller: deepLinksController,
+        );
+
+        expect(find.byType(DeepLinkPage), findsOneWidget);
+        expect(find.byType(DeepLinkListView), findsOneWidget);
+        expect(find.byType(ValidationDetailView), findsOneWidget);
       },
     );
   });
+}
+
+//TODO(hangyujin): Add more unit tests.
+class DeepLinksTestController extends DeepLinksController {
+  @override
+  Future<void> validateLinks() async {
+    displayLinkDatasNotifier.value = allLinkDatasNotifier.value;
+  }
 }