+++ packages/devtools_extensions/example/foo/packages/foo_devtools_extension/lib/src/devtools_extension_api_example.dart
@@ -9,7 +9,7 @@ import 'package:flutter/material.dart';
 
 /// This widget shows an example of how you can call public APIs exposed by
 /// [ExtensionManager].
-/// 
+///
 /// * [extensionManager.postMessageToDevTools] - how to post an arbitrary
 ///   message to DevTools, though the use for this is limited to what DevTools
 ///   is setup to handle.
@@ -27,7 +27,7 @@ class CallingDevToolsExtensionsAPIsExample extends StatelessWidget {
       crossAxisAlignment: CrossAxisAlignment.start,
       children: [
         Text(
-          '3. Example of calling DevTools extension APIs',
+          '4. Example of calling DevTools extension APIs',
           style: Theme.of(context).textTheme.titleMedium,
         ),
         const PaddedDivider.thin(),