+++ tool/lib/utils.dart
@@ -6,7 +6,6 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:args/command_runner.dart';
 import 'package:devtools_tool/model.dart';
 import 'package:io/io.dart';
 import 'package:path/path.dart' as path;
@@ -76,10 +75,13 @@ class CliCommand {
     String args, {
     bool throwOnException = true,
   }) {
+    final sdk = FlutterSdk.current;
+    if (sdk == null) {
+      throw Exception('Unable to locate a Flutter sdk.');
+    }
+
     return CliCommand._(
-      // TODO(dantup): Accept an instance of FlutterSdk instead of relying on
-      //  PATH here?
-      exe: FlutterSdk.flutterExecutableName,
+      exe: sdk.flutterToolPath,
       args: args.split(' '),
       throwOnException: throwOnException,
     );
@@ -212,12 +214,6 @@ Future<String> findRemote(
   final remoteUpstream = upstreamRemoteResult.namedGroup('remote')!;
   print('Found upstream remote.');
   return remoteUpstream;
-}
-
-extension CommandExtension on Command {
-  void logStatus(String log) {
-    print('[$name] $log');
-  }
 }
 
 extension JoinExtension on List<String> {