+++ packages/devtools_extensions/CHANGELOG.md
@@ -1,7 +1,12 @@
-## 0.0.10
+## 0.0.11
 * Migrate from `dart:html` to `package:web`.
-* Add an example of registering a service extension and calling it from the
-DevTools extension.
+
+## 0.0.10
+* Bump minimum Dart SDK version to `3.3.0-91.0.dev` and minimum Flutter SDK version to `3.17.0-0.0.pre`.
+* Add a test target to the `app_that_uses_foo` example that can also be debugged
+with the DevTools extension provided by `package:foo`.
+* Add an example of performing expression evaluations from a DevTools extension.
+* Add an example of registering a service extension and calling it from a DevTools extension.
 * Document the DevTools extension examples.
 * Add documentation to [ExtensionManager] public APIs.
 * Fix some bugs with the `build_and_copy` command for Windows.